add authorization feature

+++ SocialMediaWeb/backend/Services/ReactService.cs
@@ -11,15 +11,22 @@ namespace Backend.Services
     public class ReactService : IReactionService
     {
         public readonly BlogDbContext context;
+        public readonly IAuthService authService;
 
-        public ReactService(BlogDbContext context)
+        public ReactService(BlogDbContext context, IAuthService authService)
         {
             this.context = context;
+            this.authService = authService;
         }
 
         public ReactionResponse create(ReactionRequest reactionRequest)
         {
-            int UserId = 1;
+            int UserId = authService.GetCurrentUser().Id;
+            var post= context.Posts.FirstOrDefault(p => p.Id==reactionRequest.PostId && p.IsDeleted==false);
+            if (post == null) {
+
+                throw new Exception("post khong ton tai");
+            }
             var reactold=context.Reactions.FirstOrDefault(r=>r.UserId == UserId && r.PostId==reactionRequest.PostId);
             if (reactold == null) {
                 Reaction reactnew = new Reaction
@@ -30,20 +37,27 @@ namespace Backend.Services
                 };
                 context.Reactions.Add(reactnew);
                 context.SaveChanges();
+                reactnew.Post = post;
                 return mapToResponse(reactnew);
             }
             reactold.ReactType = reactionRequest.ReactType;
             context.SaveChanges();
+            // gán thêm data post vì add chỉ load được id của post thôi
+            reactold.Post = post;
             return mapToResponse(reactold);
         }
 
         public void Delete(int PostId)
         {
-            int UserId = 1;
+            int UserId = authService.GetCurrentUser().Id;
             var react=context.Reactions.FirstOrDefault(r=>r.UserId==UserId&&r.PostId==PostId);
             if (react == null)
             {
                 throw new Exception("người dùng chưa tương tác với bài viết này");
+            }
+            if (!authService.IsAuthor(react.UserId))
+            {
+                throw new Exception("bạn không có quyền xóa tương tác này");
             }
             context.Reactions.Remove(react);
             context.SaveChanges();
